+++ embedchain/config/AddConfig.py
@@ -1,6 +1,6 @@
 from typing import Callable, Optional
 
-from embedchain.config.BaseConfig import BaseConfig
+from embedchain.config.base_config import BaseConfig
 from embedchain.helper.json_serializable import register_deserializable
