+++ embedchain/config/__init__.py
@@ -4,7 +4,7 @@ from .AddConfig import AddConfig, ChunkerConfig
 from .apps.app_config import AppConfig
 from .apps.custom_app_config import CustomAppConfig
 from .apps.open_source_app_config import OpenSourceAppConfig
-from .BaseConfig import BaseConfig
+from .base_config import BaseConfig
 from .embedder.BaseEmbedderConfig import BaseEmbedderConfig
 from .embedder.BaseEmbedderConfig import BaseEmbedderConfig as EmbedderConfig
 from .llm.base_llm_config import BaseLlmConfig
